+++ modules/cell_tracking_module/cell_tracking.py
@@ -1,8 +1,10 @@
 import os
+import re
 import sys
 import time
 import concurrent.futures
 import logging
+import napari
 from PyQt5.QtWidgets import QLabel, QPushButton, QVBoxLayout, QWidget, QGroupBox, QApplication, QSpinBox, QFormLayout
 from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QCursor
@@ -150,8 +152,20 @@ class CellTracking(QWidget):
         self.logger = logging.getLogger(__name__)
 
     def mask_list_changed(self):
+        self.input_image.setPlaceholderText('')
+        self.input_image.setToolTip('')
         if self.mask_list.count() > 1:
             self.display_results.setChecked(False)
+        elif self.mask_list.count() == 1:
+            mask_path = self.mask_list.get_file_list()[0]
+            res = re.match('(.*)'+gf.output_suffixes['segmentation']+'.*$', os.path.basename(mask_path))
+            if res:
+                for ext in gf.imagetypes:
+                    image_path = os.path.join(os.path.dirname(mask_path), res.group(1)) + ext
+                    if os.path.isfile(image_path):
+                        self.input_image.setPlaceholderText(image_path)
+                        self.input_image.setToolTip(image_path)
+                        break
         self.display_results.setEnabled(self.mask_list.count() <= 1)
 
     def nframes_defect_changed(self, value):
@@ -214,8 +228,20 @@ class CellTracking(QWidget):
         self.nprocesses.setValue(widgets_state['nprocesses'])
 
     def submit(self):
+        # This is a temporary workaround to avoid having multiple conflicting
+        # logging to metadata and log file, which could happen when a napari
+        # window is already opened.
+        # TODO: find a better solution.
+        if napari.current_viewer():
+            self.logger.error('To avoid potential log file corruption, close all napari windows and try again.')
+            return
+
         if self.input_image.isEnabled():
             image_path = self.input_image.text()
+            if image_path == '':
+                image_path = self.input_image.placeholderText()
+            elif image_path.strip() == '':
+                image_path = ''
         else:
             image_path = ""
         mask_paths = self.mask_list.get_file_list()