+++ modules/registration_module/registration_functions.py
@@ -273,6 +273,11 @@ class EditTransformationMatrix(QWidget):
         filename = self.input_filename
         if filename != '':
             logging.getLogger(__name__).info('Saving transformation matrix to %s', filename)
+            # append to logfile
+            logfile = gf.splitext(filename)[0] + '.log'
+            with open(logfile, 'a') as f:
+                f.write(buffered_handler.get_messages())
+            # save matrix
             header = buffered_handler.get_messages()
             for x in self.tmat_metadata:
                 header += x
@@ -737,6 +742,11 @@ def registration_with_tmat(tmat, image, skip_crop, output_path, output_basename,
             ome_metadata.structured_annotations.append(CommentAnnotation(value=x, namespace="VLabApp"))
         OmeTiffWriter.save(image_cropped[0, :, :, :, :, :], registeredFilepath, ome_xml=ome_metadata)
 
+    # create logfile
+    logfile = os.path.join(output_path, output_basename+".log")
+    with open(logfile, 'w') as f:
+        f.write(buffered_handler.get_messages())
+
 
 def registration_values(image, projection_type, projection_zrange, channel_position, output_path, output_basename, registration_method, metadata, timepoint_range=None):
     """
@@ -871,6 +881,11 @@ def registration_values(image, projection_type, projection_zrange, channel_posit
         header += x
     np.savetxt(filename, transformation_matrices, fmt='%d,%d,%d,%d,%d', header=header+'x,y,keep,x_raw,y_raw', delimiter='\t')
 
+    # create logfile
+    logfile = os.path.join(output_path, output_basename+".log")
+    with open(logfile, 'w') as f:
+        f.write(buffered_handler.get_messages())
+
     return transformation_matrices
 
 
@@ -887,16 +902,10 @@ def registration_main(image_path, output_path, output_basename, channel_position
             logger.debug("creating: %s", output_path)
             os.makedirs(output_path)
 
-        logfile = os.path.join(output_path, output_basename+".log")
         logger.setLevel(logging.DEBUG)
-        logger.debug("writing log output to: %s", logfile)
-        logfile_handler = logging.FileHandler(logfile, mode='w')
-        logfile_handler.setFormatter(logging.Formatter('%(asctime)s [%(levelname)s] %(message)s'))
-        logfile_handler.setLevel(logging.INFO)
-        logger.addHandler(logfile_handler)
-        # Also save general.general_functions logger to the same file (to log information on z-projection)
-        logging.getLogger('general.general_functions').setLevel(logging.DEBUG)
-        logging.getLogger('general.general_functions').addHandler(logfile_handler)
+
+        # Log to file:
+        # saved at the end, using the content of the BufferedHandler.
 
         # Log to memory
         global buffered_handler
@@ -905,6 +914,7 @@ def registration_main(image_path, output_path, output_basename, channel_position
         buffered_handler.setLevel(logging.INFO)
         logger.addHandler(buffered_handler)
         # Also save general.general_functions logger to the same file (to log information on z-projection)
+        logging.getLogger('general.general_functions').setLevel(logging.DEBUG)
         logging.getLogger('general.general_functions').addHandler(buffered_handler)
 
         logger.info("System info:")
@@ -988,13 +998,10 @@ def alignment_main(image_path, tmat_path, output_path, output_basename, skip_cro
             logger.debug("creating: %s", output_path)
             os.makedirs(output_path)
 
-        logfile = os.path.join(output_path, output_basename+".log")
         logger.setLevel(logging.DEBUG)
-        logger.debug("writing log output to: %s", logfile)
-        logfile_handler = logging.FileHandler(logfile, mode='w')
-        logfile_handler.setFormatter(logging.Formatter('%(asctime)s [%(levelname)s] %(message)s'))
-        logfile_handler.setLevel(logging.INFO)
-        logger.addHandler(logfile_handler)
+
+        # Log to file:
+        # saved at the end, using the content of the BufferedHandler.
 
         # Log to memory
         global buffered_handler
@@ -1071,94 +1078,16 @@ def alignment_main(image_path, tmat_path, output_path, output_basename, skip_cro
 ################################################################
 
 
-def edit_main(reference_matrix_path, range_start, range_end):
-    try:
-        log_path = gf.splitext(reference_matrix_path)[0] + '.log'
-
-        # Setup logging to file in output_path
-        logger = logging.getLogger(__name__)
-        logger.info("REGISTRATION MODULE (editing)")
-
-        logfile = os.path.join(log_path)
-        logger.setLevel(logging.DEBUG)
-        logger.debug("writing log output to: %s", logfile)
-        logfile_handler = logging.FileHandler(logfile, mode='a')
-        logfile_handler.setFormatter(logging.Formatter('%(asctime)s [%(levelname)s] %(message)s'))
-        logfile_handler.setLevel(logging.INFO)
-        logger.addHandler(logfile_handler)
-
-        # Log to memory
-        global buffered_handler
-        buffered_handler = gf.BufferedHandler()
-        buffered_handler.setFormatter(logging.Formatter('%(asctime)s (VLabApp - registration module) [%(levelname)s] %(message)s'))
-        buffered_handler.setLevel(logging.INFO)
-        logger.addHandler(buffered_handler)
-
-        logger.info("System info:")
-        logger.info("- platform: %s", platform())
-        logger.info("- python version: %s", python_version())
-        logger.info("- VLabApp version: %s", vlabapp_version)
-        logger.info("- numpy version: %s", np.__version__)
-        logger.info("- pystackreg version: %s", StackReg_version)
-        logger.info("- opencv version: %s", cv.__version__)
-        logger.info("- skimage version: %s", skimage_version)
-
-        logger.info("Input transformation matrix path: %s", reference_matrix_path)
-
-        # Load the transformation matrix
-        logger.debug("loading: %s", reference_matrix_path)
-        tmat, tmat_metadata = read_transformation_matrix(reference_matrix_path)
-
-        if range_start >= tmat.shape[0]:
-            logger.error('Invalid timepoint range')
-            remove_all_log_handlers()
-            raise ValueError('Invalid timepoint range')
-
-        # Update transformation matrix
-        logger.info("Editing transformation matrix (start=%s, end=%s)", range_start, range_end)
-        tmat[:, 2] = 0
-        tmat[range_start:(range_end+1), 2] = 1
-        tmat[:, 0] = tmat[:, 3] - tmat[range_start, 3]
-        tmat[:, 1] = tmat[:, 4] - tmat[range_start, 4]
-        tmat[:range_start, 0] = tmat[range_start, 0]
-        tmat[:range_start, 1] = tmat[range_start, 1]
-        if range_end < tmat.shape[0]:
-            tmat[range_end:, 0] = tmat[range_end, 0]
-            tmat[range_end:, 1] = tmat[range_end, 1]
-
-        # Save the new matrix
-        logger.info("Saving transformation matrix to %s", reference_matrix_path)
-        header = buffered_handler.get_messages()
-        for x in tmat_metadata:
-            header += x
-        np.savetxt(reference_matrix_path, tmat, fmt='%d,%d,%d,%d,%d', header=header+'x,y,keep,x_raw,y_raw', delimiter='\t')
-
-        remove_all_log_handlers()
-
-    except Exception:
-        # Remove all handlers for this module
-        remove_all_log_handlers()
-        raise
-
-
-################################################################
-
-
 def manual_edit_main(image_path, matrix_path):
     try:
-        log_path = gf.splitext(matrix_path)[0] + '.log'
-
         # Setup logging to file in output_path
         logger = logging.getLogger(__name__)
         logger.info("REGISTRATION MODULE (manual editing)")
 
-        logfile = os.path.join(log_path)
         logger.setLevel(logging.DEBUG)
-        logger.debug("writing log output to: %s", logfile)
-        logfile_handler = logging.FileHandler(logfile, mode='a')
-        logfile_handler.setFormatter(logging.Formatter('%(asctime)s [%(levelname)s] %(message)s'))
-        logfile_handler.setLevel(logging.INFO)
-        logger.addHandler(logfile_handler)
+
+        # Log to file:
+        # saved at the end, using the content of the BufferedHandler.
 
         # Log to memory
         global buffered_handler